refactor: Move `*GroupBy` + friends
