+++ narwhals/_plan/arrow/group_by.py
@@ -9,8 +9,8 @@ from narwhals._plan import expressions as ir
 from narwhals._plan._guards import is_agg_expr, is_function_expr
 from narwhals._plan.arrow import acero, functions as fn, options
 from narwhals._plan.common import dispatch_method_name, temp
+from narwhals._plan.compliant.group_by import EagerDataFrameGroupBy
 from narwhals._plan.expressions import aggregation as agg
-from narwhals._plan.protocols import EagerDataFrameGroupBy
 from narwhals._utils import Implementation
 from narwhals.exceptions import InvalidOperationError
